Changed program to read the file char by char instead of one big string.
I should note that (lol) currently the program doesn't actually do the capitalization. If you uncomment lines 22 and 23 in Program.cs, the program should theoretically check to see if the character is a lowercase letter and then add enough value to it to make it uppercase. That is to say:

Uppercase = Lowercase + 'A' - 'a'

But there's an error where C# doesn't like the addition of a char to anything. The error specifically says 'Cannot implictly convert char to int' in line 23.