Program works correctly.
Program reads characters continuously from a file. If the character is a lowercase letter then it will be converted to uppercase. The character (regardless of it was converted or not) will be sent to the output file.

I changed some of the variable names so they're easier to understand.
I also moved the comments along a line on the right for readability.