You must now use the command line to enter file name.
Do this using the Developer Command Prompt from Visual Studio 2013 tools folder.
I also changed the variable names to be all lowercase.
And I added some failsafes in case you don't actually enter a file name or if the file doesn't exist.
Also changed an if statement so the integer representing the char we're reading doesn't get cast to a char for the comparison.